traffic_project_smooth/traffic_project/requirements.txt
opencv-python
easyocr
openpyxl
fpdf2
pillow
numpy
